+ creadas las clases CExerciseEncoding y CEncoderJeringozo + creada la clase abstracta CExercise + renombradas las clases FranciscoGinzo, FranciscoGinzoTest, ExerciseGitSetUp, ExerciseStudentCreation y ExerciseStudentCreationTest + extendidas las clases CExerciseEncoding, CExerciseGitSetUp y CExerciseStudentCreation de CExercise
