+++ ejercicios/src/ar/edu/ub/pcsw2016/fginzo/CExerciseGitSetUp.java
@@ -1,0 +1,20 @@
+package ar.edu.ub.pcsw2016.fginzo;
+
+import ar.edu.ub.pcsw2016.IExercise;
+
+public class CExerciseGitSetUp extends CExercise implements IExercise {
+
+    private ID id;
+
+    /*
+    Constructor
+     */
+    public CExerciseGitSetUp(ID id) {
+        this.id = id;
+    }
+
+    @Override
+    public void run() {
+        System.out.println("fginzo");
+    }
+}
