+++ ejercicios/src/ar/edu/ub/pcsw2016/fginzo/CExerciseStudentCreation.java
@@ -1,0 +1,20 @@
+package ar.edu.ub.pcsw2016.fginzo;
+
+import ar.edu.ub.pcsw2016.IExercise;
+
+public class CExerciseStudentCreation extends CExercise implements IExercise {
+
+    private IExercise.ID id;
+
+    /*
+    Constructor
+     */
+    public CExerciseStudentCreation(IExercise.ID id) {
+        this.id = id;
+    }
+
+    @Override
+    public void run() {
+        System.out.println(new CStudentFranciscoGinzo().name());
+    }
+}
