--- ejercicios/src/ar/edu/ub/pcsw2016/fginzo/ExerciseStudentCreation.java
@@ -1,55 +1,0 @@
-package ar.edu.ub.pcsw2016.fginzo;
-
-import ar.edu.ub.pcsw2016.IExercise;
-
-public class ExerciseStudentCreation implements IExercise {
-
-    private float grade;
-    private IExercise.ID id;
-
-    /*
-    Constructor
-     */
-    public ExerciseStudentCreation (IExercise.ID id) {
-        this.id = id;
-    }
-
-    /*
-    Getter del ID del ejercicio
-     */
-    @Override
-    public IExercise.ID id() {
-        return this.id;
-    }
-
-    @Override
-    public void run() {
-        System.out.println(new FranciscoGinzo().name());
-    }
-
-    /*
-    Setter de la nota del ejercicio sólo en un rango predefinido
-     */
-    @Override
-    public void setGrade(float grade) {
-        if (grade >= 0 && grade <= 10) {
-            this.grade = grade;
-        }
-        else {
-            System.out.println("La nota ingresada no corresponde al sistema de calificación (0 a 10 ambos inclusive).");
-        }
-    }
-
-    /*
-    Getter de la nota asignada al ejercicio
-     */
-    @Override
-    public float grade() {
-        return grade;
-    }
-
-    public String toString() {
-        return this.id() + " ";
-    }
-
-}
